Generalise manifest format
