+++ upki-mirror/src/bin/intermediates.rs
@@ -10,7 +10,7 @@ use eyre::{Context, Report, anyhow};
 use rustls_pki_types::CertificateDer;
 use rustls_pki_types::pem::PemObject;
 use serde::Deserialize;
-use upki::revocation::{Manifest, ManifestFile};
+use upki::data::{Manifest, ManifestFile};
 
 #[tokio::main(flavor = "current_thread")]
 async fn main() -> Result<(), Report> {
