+++ upki/src/lib.rs
@@ -207,3 +207,6 @@ pub mod revocation;
 
 /// Fetching intermediate certificates to assist chain building.
 pub mod intermediates;
+
+/// Data storage.
+pub mod data;
